moved radar to home page
