+++ src/pages/Roadmap/index.module.css
@@ -58,45 +58,6 @@
   margin-left: auto;
 }
 
-/* just for mvp */
-.todo_div {
-  border: 6px #0198ba;
-  border-style: solid;
-  border-radius: 50px;
-  margin-top: 10vw;
-  margin-left: 2vw;
-  margin-right: 2vw;
-  padding-left: 3vw;
-  padding-bottom: 10vw;
-}
-
-.radar {
-  font-family: "Poppins";
-  font-weight: 900;
-  color: black;
-  font-size: 7vw;
-  margin-top: 3vw;
-}
-
-.priority {
-  font-family: "Poppins";
-  font-weight: bolder;
-  color: #8b5674;
-  font-size: 5vw;
-  margin-top: 6vw;
-}
-
-.priority.else {
-  color: #e3d150;
-}
-
-.list {
-  font-family: "Poppins";
-  color: black;
-  font-weight: 600;
-  font-size: 3vw;
-}
-
 .csvButton {
   margin-bottom: 2vh;
 }
